requirements.txt
Django==4.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
pika==1.3.2
Pillow==10.1.0
python-decouple==3.8
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0

